% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camostat_model_sum_stats.R
\name{calc_r_camostat_stages}
\alias{calc_r_camostat_stages}
\title{calculates r for the TEIV model with Erlang-distributed latent and infectious periods, and different cell types}
\usage{
calc_r_camostat_stages(params, warn_if_neg = FALSE)
}
\arguments{
\item{params}{a list or named numeric vector with the parameter values for
beta: infectivity in inverse units of V (in dT/dt)
T_0: initial number of target cells in units of T
p: production rate of infectious virus from cells
c_inf: decay rate of infectious virus
k1: rate from latent to infectious
delta: decay rate of infectious cells
n_L: shape parameter for distribution of latent period
n_I: shape parameter for distribution of infectious period
beta_inf: infectivity in inverse units of T (in dV/dt)
if loss of virus due to enry into target cells is ignored, beta_inf = 0}
}
\value{
r: numeric vector of length 1
}
\description{
calculates r for the TEIV model with Erlang-distributed latent and infectious periods, and different cell types
}
