% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camostat_model_calc_sol.R
\name{calc_sol_add_stages}
\alias{calc_sol_add_stages}
\title{for a set of parameters without stages, add stages and recalculate model solution}
\usage{
calc_sol_add_stages(model, pars, n_L, n_I)
}
\arguments{
\item{model}{string:"full", "endosomal_only", or "tmprss2_only"}

\item{pars}{vector of parameter values}

\item{n_L}{number of latent stages}

\item{n_I}{number of infectious stages}
}
\value{
data frame outputted by odin
}
\description{
for a set of parameters without stages, add stages and recalculate model solution
}
