% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camostat_model_calc_sol.R
\name{calc_sol_chain}
\alias{calc_sol_chain}
\title{calculate solution to ODEs for each pathway, and extract median, 95\\% CI and max LL}
\usage{
calc_sol_chain(filename, model, solving_time_max = 72)
}
\arguments{
\item{filename}{.RData file produced by lazymcmc}

\item{model}{string: "full", "endosomal_only", "tmprss2_only"}

\item{solving_time_max}{time to solve until.   Default = 72 hpi.}
}
\value{
tibble with columns model, t, V, T1... and probs, which can take the values 0.025, 0.5, 0.975 or "max_LL"
}
\description{
calculate solution to ODEs for each pathway, and extract median, 95\\% CI and max LL
}
