% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camostat_model_posterior.R
\name{CREATE_POSTERIOR_FUNC_fn_camostat_amphoB}
\alias{CREATE_POSTERIOR_FUNC_fn_camostat_amphoB}
\title{creates a closure to calculate the likelihood and model predictions for given
parameter values}
\usage{
CREATE_POSTERIOR_FUNC_fn_camostat_amphoB(
  parTab,
  data_df,
  PRIOR_FUNC,
  camostat_vec,
  amphoB_vec
)
}
\value{
closure to calculate likelihood and model predictions for given
parameter values
}
\description{
creates a closure to calculate the likelihood and model predictions for given
parameter values
}
