% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camostat_model_sum_stats.R
\name{gen_summary_statistics_fn_camostat}
\alias{gen_summary_statistics_fn_camostat}
\title{create closure to calculate summary statistics and provide bounds for them}
\usage{
gen_summary_statistics_fn_camostat(parTab)
}
\arguments{
\item{parTab}{data frame}
}
\value{
a list with elements
"calc_summary": closure to calculate
}
\description{
create closure to calculate summary statistics and provide bounds for them
}
