% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camostat_model_other.R
\name{get_strain_names}
\alias{get_strain_names}
\title{determine strain names from parameter names}
\usage{
get_strain_names(par_names)
}
\arguments{
\item{par_names}{character vector: parameter names}
}
\value{
character vector string names
}
\description{
determine strain names from parameter names
}
