% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camostat_model_other.R
\name{plot_dynamics_camostat}
\alias{plot_dynamics_camostat}
\title{plot viral dynamics}
\usage{
plot_dynamics_camostat(parTab, data_df, chain)
}
\arguments{
\item{parTab}{data frame}

\item{chain}{data table of samples from MCMC chain}
}
\value{
list of ggplot objects: plots of data_df, 95\% credible intervals and
and sampled trajectories for viral loads
}
\description{
plot viral dynamics
}
