% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_Calu3_data}
\alias{read_Calu3_data}
\title{read Calu 3 data with different drugs}
\usage{
read_Calu3_data(strain, from_raw = FALSE, PCR = TRUE)
}
\description{
read Calu 3 data with different drugs
}
