% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_camostat_data}
\alias{read_camostat_data}
\title{read data with and without camostat}
\usage{
read_camostat_data(strain1, from_raw = FALSE, PCR = TRUE)
}
\arguments{
\item{from_raw}{logical, whether to read from raw data}

\item{PCR}{logical, whetehr to include PCR data}

\item{strain}{WT, DeltaCS, delta or omicron}
}
\value{
a data frame
}
\description{
read data with and without camostat
}
