% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_plaque_peacock}
\alias{read_plaque_peacock}
\title{read plaque assay data from Peacock et al. (2021) Nat Microbiol 10.1038/s41564-021-00908-w}
\usage{
read_plaque_peacock()
}
\description{
read plaque assay data from Peacock et al. (2021) Nat Microbiol 10.1038/s41564-021-00908-w
}
