% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camostat_model_run.R
\name{run_exp_camostat_amphoB}
\alias{run_exp_camostat_amphoB}
\title{estimate parameters for the different models in the study}
\usage{
run_exp_camostat_amphoB(
  save_dir,
  strain,
  camostat_vec,
  amphoB_vec,
  Calu3 = FALSE,
  PCR = FALSE,
  mvr = TRUE,
  length_run = 2,
  run_flag = TRUE
)
}
\arguments{
\item{save_dir}{string ending in "/": directory to save results in}

\item{strain}{character vector.  names of strains.}

\item{camostat_vec}{logical vector of length n, where n is the number of conditions.  Each value
indicates whether camostat was used in the condition.}

\item{amphoB_vec}{logical vector of length n, where n is the number of conditions.  Each value
indicates whether amphoB was used in the condition.}

\item{Calu3}{logical. if TRUE, fit model to Calu3 data, if FALSE, use HAE data}

\item{PCR}{logical.  whether to include PCR data (TRUE) or only plaque data (FALSE)}

\item{mvr}{logical.  If TRUE, use multivariate proposal, if FALSE, use univariate proposal.}

\item{length_run}{integer.  If length == 1, short run (for testing purposes); if length == 2, normal run; if length = 3, extra long run.
otherwise, run until convergence}

\item{run_flag}{logial.  if TRUE, run and postprocess; if FALSE, postprocess only}
}
\value{
NULL (results saved to file)
}
\description{
estimate parameters for the different models in the study
}
