% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camostat_model_run.R
\name{run_exp_camostat_fit_subset_pars}
\alias{run_exp_camostat_fit_subset_pars}
\title{estimate prob_infect_tmprss2 and prob_infect_camostat, fix other parameters}
\usage{
run_exp_camostat_fit_subset_pars(
  save_dir,
  strain,
  old_values,
  old_parTab,
  subset_pars = c("prob_infection_tmprss2", "prob_infection_endosomal"),
  mvr = TRUE,
  length_run = 2,
  run_flag = TRUE
)
}
\arguments{
\item{save_dir}{string ending in "/": directory to save results in}

\item{strain}{character vector.  names of strains.}

\item{old_values}{named numeric vector of fixed parameter values}

\item{old_parTab}{old parameter table}

\item{subset_pars}{character vector.  Subset of parameters to be fitted.}

\item{mvr}{logical. if TRUE, use multivariate proposal; if FALSE, use univariate proposal}

\item{length_run}{integer.  If length == 1, short run (for testing purposes); if length == 2, normal run; if length = 3, extra long run.
otherwise, run until convergence}

\item{run_flag}{logial.  if TRUE, run and postprocess; if FALSE, postprocess only}
}
\value{
NULL (results saved to file)
}
\description{
estimate prob_infect_tmprss2 and prob_infect_camostat, fix other parameters
}
