% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camostat_model_other.R
\name{solve_from_posterior_preincub}
\alias{solve_from_posterior_preincub}
\title{solve preincubation viral dynamics for one iteration of chain}
\usage{
solve_from_posterior_preincub(subchain)
}
\arguments{
\item{subchain}{data table of samples from MCMC chain, should have 1 row only}
}
\value{
data frame
}
\description{
solve preincubation viral dynamics for one iteration of chain
}
