% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camostat_model_specify_pars.R
\name{specify_camostat_parameters_fn}
\alias{specify_camostat_parameters_fn}
\title{make parameter table}
\usage{
specify_camostat_parameters_fn(
  n_L = 1,
  n_I = 1,
  different_eclipse = FALSE,
  assay = "plaque",
  PCR = FALSE,
  reparameterise = FALSE,
  k1_endosomal_limit = FALSE,
  strain_names
)
}
\arguments{
\item{k1_endosomal_limit}{logical. if TRUE, 0 <= k1_endosomal/k1_tmprss2 <= 1}

\item{strain_names}{vector of strings.  Names of virus strains.}

\item{n_L:}{integer. number of latent stages}

\item{n_I:}{integer.  number of infectious stages}

\item{assay:}{string.  "plaque" or "tcid50" (only affects observation threshold)}

\item{reparameterise:}{logical.  if FALSE, use beta, p etc, if TRUE, use probability of infection (beta T_0) / c, burst size, delta, c}
}
\value{
data_df frame with columns
values: default parameter value
names: internal name for parameter
fixed:  if 0, fit parameter, otherwise fix parameter
lower_bound: lower bound of parameter for fitting
upper_bound: upper bound of parameter for fitting
steps: width of proposal distribution in parameter space
names_plot: character vector which which to label parameter distribution plot
}
\description{
make parameter table
}
