% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camostat_model_specify_pars.R
\name{specify_camostat_parameters_fn_fit_subset_pars}
\alias{specify_camostat_parameters_fn_fit_subset_pars}
\title{make parameter table where only some parameters are fitted
specify fixed parameter values}
\usage{
specify_camostat_parameters_fn_fit_subset_pars(
  old_values,
  old_parTab,
  subset_pars
)
}
\arguments{
\item{old_values}{named numeric vector of fixed parameter values}

\item{old_parTab}{old parameter table}

\item{subset_pars}{character vector.  Subset of parameters to be fitted.}
}
\value{
data frame: new parameter table
}
\description{
make parameter table where only some parameters are fitted
specify fixed parameter values
}
