% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camostat_model_other.R
\name{transform_pars_wrapper_tmprss2_only}
\alias{transform_pars_wrapper_tmprss2_only}
\title{create a closure which transforms an unnamed numeric vector into
lists of named parameters to calculate the likelihood}
\usage{
transform_pars_wrapper_tmprss2_only(parTab)
}
\arguments{
\item{parTab}{data frame containing priors}
}
\value{
a function which transforms an unnamed numeric vector into
lists of named parameters to calculate the likelihood
}
\description{
create a closure which transforms an unnamed numeric vector into
lists of named parameters to calculate the likelihood
}
